Revert "Re-Revert "EVG-5161 don't restart display tasks until all finished (#1512)""
This reverts commit f5c2d1a9798aa624d1b7789b4c276d52c27c850d.
