+++ model/task_lifecycle.go
@@ -2,6 +2,7 @@ package model
 
 import (
 	"fmt"
+	"sort"
 	"time"
 
 	"github.com/evergreen-ci/evergreen"
@@ -16,6 +17,7 @@ import (
 	"github.com/mongodb/grip/message"
 	"github.com/pkg/errors"
 	mgo "gopkg.in/mgo.v2"
+	"gopkg.in/mgo.v2/bson"
 )
 
 type StatusChanges struct {
@@ -89,7 +91,7 @@ func SetActiveState(taskId string, caller string, active bool) error {
 	}
 
 	if t.IsPartOfDisplay() {
-		return updateDisplayTask(t)
+		return updateDisplayTaskAndCache(t)
 	}
 
 	return errors.WithStack(build.SetCachedTaskActivated(t.BuildId, taskId, active))
@@ -228,9 +230,6 @@ func TryResetTask(taskId, user, origin string, detail *apimodels.TaskEndDetail) 
 		event.LogTaskRestarted(t.Id, t.Execution, origin)
 	}
 
-	if t.DisplayOnly {
-		return t.UpdateDisplayTask()
-	}
 	return errors.WithStack(err)
 }
 
@@ -409,7 +408,7 @@ func MarkEnd(t *task.Task, caller string, finishTime time.Time, detail *apimodel
 	event.LogTaskFinished(t.Id, t.Execution, t.HostId, status)
 
 	if t.IsPartOfDisplay() {
-		if err = t.DisplayTask.UpdateDisplayTask(); err != nil {
+		if err = UpdateDisplayTask(t.DisplayTask); err != nil {
 			return err
 		}
 		if err = build.UpdateCachedTask(t.DisplayTask.BuildId, t.DisplayTask.Id, t.DisplayTask.Status, t.TimeTaken); err != nil {
@@ -534,7 +533,7 @@ func UpdateBuildAndVersionStatusForTask(taskId string, updates *StatusChanges) e
 				return err
 			}
 			if displayTask != nil {
-				err = displayTask.UpdateDisplayTask()
+				err = UpdateDisplayTask(displayTask)
 				if err != nil {
 					return err
 				}
@@ -700,7 +699,7 @@ func MarkStart(t *task.Task, updates *StatusChanges) error {
 	}
 
 	if t.IsPartOfDisplay() {
-		return updateDisplayTask(t)
+		return updateDisplayTaskAndCache(t)
 	}
 
 	// update the cached version of the task, in its build document
@@ -716,7 +715,7 @@ func MarkTaskUndispatched(t *task.Task) error {
 	event.LogTaskUndispatched(t.Id, t.Execution, t.HostId)
 
 	if t.IsPartOfDisplay() {
-		return updateDisplayTask(t)
+		return updateDisplayTaskAndCache(t)
 	}
 
 	// update the cached version of the task in its related build document
@@ -736,7 +735,7 @@ func MarkTaskDispatched(t *task.Task, hostId, distroId string) error {
 	event.LogTaskDispatched(t.Id, t.Execution, hostId)
 
 	if t.IsPartOfDisplay() {
-		return updateDisplayTask(t)
+		return updateDisplayTaskAndCache(t)
 	}
 
 	// update the cached version of the task in its related build document
@@ -746,8 +745,8 @@ func MarkTaskDispatched(t *task.Task, hostId, distroId string) error {
 	return nil
 }
 
-func updateDisplayTask(t *task.Task) error {
-	err := t.DisplayTask.UpdateDisplayTask()
+func updateDisplayTaskAndCache(t *task.Task) error {
+	err := UpdateDisplayTask(t.DisplayTask)
 	if err != nil {
 		return errors.Wrap(err, "error updating display task")
 	}
@@ -880,11 +879,112 @@ func ClearAndResetStrandedTask(h *host.Host) error {
 	}
 
 	if time.Since(t.StartTime) < task.UnschedulableThreshold {
-		return errors.Wrap(TryResetTask(t.Id, "mci", evergreen.MonitorPackage, &apimodels.TaskEndDetail{
+		detail := &apimodels.TaskEndDetail{
 			Status: evergreen.TaskFailed,
 			Type:   "system",
-		}), "problem resetting task")
+		}
+		if t.IsPartOfDisplay() {
+			return t.DisplayTask.SetResetWhenFinished(detail)
+		}
+		return errors.Wrap(TryResetTask(t.Id, "mci", evergreen.MonitorPackage, detail), "problem resetting task")
 	}
 
+	return nil
+}
+
+func UpdateDisplayTask(t *task.Task) error {
+	if !t.DisplayOnly {
+		return fmt.Errorf("%s is not a display task", t.Id)
+	}
+
+	statuses := []string{}
+	var timeTaken time.Duration
+	var status string
+	wasFinished := t.IsFinished()
+	execTasks, err := task.Find(task.ByIds(t.ExecutionTasks))
+	if err != nil {
+		return errors.Wrap(err, "error retrieving execution tasks")
+	}
+	hasFinishedTasks := false
+	hasUnfinishedTasks := false
+	startTime := time.Unix(1<<62, 0)
+	endTime := util.ZeroTime
+	for _, execTask := range execTasks {
+		// if any of the execution tasks are scheduled, the display task is too
+		if execTask.Activated {
+			t.Activated = true
+		}
+
+		if execTask.IsFinished() {
+			hasFinishedTasks = true
+		} else if execTask.IsDispatchable() {
+			hasUnfinishedTasks = true
+		}
+
+		// the display task's status will be the highest priority of its exec tasks
+		statuses = append(statuses, execTask.ResultStatus())
+
+		// add up the duration of the execution tasks as the cumulative time taken
+		timeTaken += execTask.TimeTaken
+
+		// set the start/end time of the display task as the earliest/latest task
+		if execTask.StartTime.Before(startTime) {
+			startTime = execTask.StartTime
+		}
+		if execTask.FinishTime.After(endTime) {
+			endTime = execTask.FinishTime
+		}
+	}
+
+	if hasFinishedTasks && hasUnfinishedTasks {
+		// if the display task has a mix of finished and unfinished tasks, the status
+		// will be "started"
+		status = evergreen.TaskStarted
+	} else if len(statuses) > 0 {
+		// the status of the display task will be the status of its constituent task
+		// that is logically the most exclusive
+		sort.Sort(task.ByPriority(statuses))
+		status = statuses[0]
+	}
+
+	grip.InfoWhen(status == evergreen.TaskUndispatched && t.DispatchTime != util.ZeroTime, message.Fields{
+		"lookhere":                      "evg-3345",
+		"message":                       "update display tasks",
+		"task_id":                       t.Id,
+		"status":                        status,
+		"dispatch_time_is_go_zero_time": t.DispatchTime.IsZero(),
+	})
+
+	update := bson.M{
+		task.StatusKey:    status,
+		task.ActivatedKey: t.Activated,
+		task.TimeTakenKey: timeTaken,
+	}
+	if startTime != time.Unix(1<<62, 0) {
+		update[task.StartTimeKey] = startTime
+	}
+	if endTime != util.ZeroTime && !hasUnfinishedTasks {
+		update[task.FinishTimeKey] = endTime
+	}
+
+	err = task.UpdateOne(
+		bson.M{
+			task.IdKey: t.Id,
+		},
+		bson.M{
+			"$set": update,
+		})
+	if err != nil {
+		return errors.Wrap(err, "error updating display task")
+	}
+
+	t.Status = status
+	t.TimeTaken = timeTaken
+	if !wasFinished && t.IsFinished() {
+		event.LogDisplayTaskFinished(t.Id, t.Execution, t.Status)
+		if t.ResetWhenFinished {
+			return errors.Wrap(TryResetTask(t.Id, evergreen.User, evergreen.User, &t.Details), "error resetting display task")
+		}
+	}
 	return nil
 }
