+++ model/task_lifecycle_test.go
@@ -9,6 +9,7 @@ import (
 	"github.com/evergreen-ci/evergreen/apimodels"
 	"github.com/evergreen-ci/evergreen/db"
 	"github.com/evergreen-ci/evergreen/model/build"
+	"github.com/evergreen-ci/evergreen/model/event"
 	"github.com/evergreen-ci/evergreen/model/task"
 	"github.com/evergreen-ci/evergreen/model/version"
 	"github.com/evergreen-ci/evergreen/testutil"
@@ -1036,7 +1037,7 @@ func TestTryResetTask(t *testing.T) {
 		So(t1FromDb.Status, ShouldEqual, evergreen.TaskUndispatched)
 		dtFromDb, err := task.FindOne(task.ById(dt.Id))
 		So(err, ShouldBeNil)
-		So(dtFromDb.Status, ShouldEqual, evergreen.TaskUnstarted)
+		So(dtFromDb.Status, ShouldEqual, evergreen.TaskUndispatched)
 		dbBuild, err := build.FindOne(build.ById(b.Id))
 		So(err, ShouldBeNil)
 		So(dbBuild.Tasks[0].Status, ShouldEqual, evergreen.TaskUndispatched)
@@ -1935,4 +1936,166 @@ func TestMarkEndRequiresAllTasksToFinishToUpdateBuildStatusWithCompileTask(t *te
 	assert.True(complete)
 	assert.NoError(err)
 	assert.True(b.IsFinished())
+}
+
+func TestDisplayTaskUpdates(t *testing.T) {
+	testutil.HandleTestingErr(db.ClearCollections(task.Collection, event.AllLogCollection), t, "error clearing collection")
+	assert := assert.New(t)
+	dt := task.Task{
+		Id:          "dt",
+		DisplayOnly: true,
+		Status:      evergreen.TaskUndispatched,
+		Activated:   false,
+		ExecutionTasks: []string{
+			"task1",
+			"task2",
+			"task3",
+			"task4",
+		},
+	}
+	assert.NoError(dt.Insert())
+	dt2 := task.Task{
+		Id:          "dt2",
+		DisplayOnly: true,
+		Status:      evergreen.TaskUndispatched,
+		Activated:   false,
+		ExecutionTasks: []string{
+			"task5",
+			"task6",
+		},
+	}
+	assert.NoError(dt2.Insert())
+	task1 := task.Task{
+		Id:         "task1",
+		Status:     evergreen.TaskFailed,
+		TimeTaken:  3 * time.Minute,
+		StartTime:  time.Date(2000, 0, 0, 1, 1, 1, 0, time.Local),
+		FinishTime: time.Date(2000, 0, 0, 1, 9, 1, 0, time.Local),
+	}
+	assert.NoError(task1.Insert())
+	task2 := task.Task{
+		Id:         "task2",
+		Status:     evergreen.TaskSucceeded,
+		TimeTaken:  2 * time.Minute,
+		StartTime:  time.Date(2000, 0, 0, 0, 30, 0, 0, time.Local), // this should end up as the start time for dt1
+		FinishTime: time.Date(2000, 0, 0, 1, 0, 5, 0, time.Local),
+	}
+	assert.NoError(task2.Insert())
+	task3 := task.Task{
+		Id:         "task3",
+		Activated:  true,
+		Status:     evergreen.TaskSystemUnresponse,
+		TimeTaken:  5 * time.Minute,
+		StartTime:  time.Date(2000, 0, 0, 0, 44, 0, 0, time.Local),
+		FinishTime: time.Date(2000, 0, 0, 1, 0, 1, 0, time.Local),
+	}
+	assert.NoError(task3.Insert())
+	task4 := task.Task{
+		Id:         "task4",
+		Activated:  true,
+		Status:     evergreen.TaskSystemUnresponse,
+		TimeTaken:  1 * time.Minute,
+		StartTime:  time.Date(2000, 0, 0, 1, 0, 20, 0, time.Local),
+		FinishTime: time.Date(2000, 0, 0, 1, 22, 0, 0, time.Local), // this should end up as the end time for dt1
+	}
+	assert.NoError(task4.Insert())
+	task5 := task.Task{
+		Id:        "task5",
+		Activated: true,
+		Status:    evergreen.TaskUndispatched,
+	}
+	assert.NoError(task5.Insert())
+	task6 := task.Task{
+		Id:        "task6",
+		Activated: true,
+		Status:    evergreen.TaskSucceeded,
+	}
+	assert.NoError(task6.Insert())
+
+	// test that updating the status + activated from execution tasks works
+	assert.NoError(UpdateDisplayTask(&dt))
+	dbTask, err := task.FindOne(task.ById(dt.Id))
+	assert.NoError(err)
+	assert.NotNil(dbTask)
+	assert.Equal(evergreen.TaskFailed, dbTask.Status)
+	assert.True(dbTask.Activated)
+	assert.Equal(11*time.Minute, dbTask.TimeTaken)
+	assert.Equal(task2.StartTime, dbTask.StartTime)
+	assert.Equal(task4.FinishTime, dbTask.FinishTime)
+
+	// test that you can't update an execution task
+	assert.Error(UpdateDisplayTask(&task1))
+
+	// test that a display task with a finished + unstarted task is "scheduled"
+	assert.NoError(UpdateDisplayTask(&dt2))
+	dbTask, err = task.FindOne(task.ById(dt2.Id))
+	assert.NoError(err)
+	assert.NotNil(dbTask)
+	assert.Equal(evergreen.TaskStarted, dbTask.Status)
+
+	// check that the updates above logged an event for the first one
+	events, err := event.Find(event.AllLogCollection, event.TaskEventsForId(dt.Id))
+	assert.NoError(err)
+	assert.Len(events, 1)
+	events, err = event.Find(event.AllLogCollection, event.TaskEventsForId(dt2.Id))
+	assert.NoError(err)
+	assert.Len(events, 0)
+}
+
+func TestDisplayTaskDelayedRestart(t *testing.T) {
+	testutil.HandleTestingErr(db.ClearCollections(task.Collection, task.OldCollection, build.Collection), t, "error clearing collection")
+	assert := assert.New(t)
+	dt := task.Task{
+		Id:          "dt",
+		DisplayOnly: true,
+		Status:      evergreen.TaskStarted,
+		Activated:   true,
+		BuildId:     "b",
+		ExecutionTasks: []string{
+			"task1",
+			"task2",
+		},
+	}
+	assert.NoError(dt.Insert())
+	task1 := task.Task{
+		Id:      "task1",
+		BuildId: "b",
+		Status:  evergreen.TaskSucceeded,
+	}
+	assert.NoError(task1.Insert())
+	task2 := task.Task{
+		Id:      "task2",
+		BuildId: "b",
+		Status:  evergreen.TaskSucceeded,
+	}
+	assert.NoError(task2.Insert())
+	b := build.Build{
+		Id: "b",
+		Tasks: []build.TaskCache{
+			{Id: "dt", Status: evergreen.TaskStarted, Activated: true},
+		},
+	}
+	assert.NoError(b.Insert())
+
+	// request that the task restarts when it's done
+	detail := &apimodels.TaskEndDetail{
+		Description: "foo",
+		Type:        evergreen.CommandTypeTest,
+		Status:      evergreen.TaskSucceeded,
+	}
+	assert.NoError(dt.SetResetWhenFinished(detail))
+	dbTask, err := task.FindOne(task.ById(dt.Id))
+	assert.NoError(err)
+	assert.Equal(detail.Description, dbTask.Details.Description)
+	assert.True(dbTask.ResetWhenFinished)
+	assert.Equal(evergreen.TaskStarted, dbTask.Status)
+
+	// end the final task so that it restarts
+	assert.NoError(UpdateDisplayTask(&dt))
+	dbTask, err = task.FindOne(task.ById(dt.Id))
+	assert.NoError(err)
+	assert.Equal(evergreen.TaskUndispatched, dbTask.Status)
+	oldTask, err := task.FindOneOld(task.ById("dt_0"))
+	assert.NoError(err)
+	assert.Equal(detail.Description, oldTask.Details.Description)
 }
