+++ units/task_monitor_execution_timeout.go
@@ -80,7 +80,6 @@ func (j *taskExecutionTimeoutJob) Run(ctx context.Context) {
 		return
 	}
 
-	displayTasksRestarted := map[string]bool{}
 	for _, task := range tasks {
 		msg := message.Fields{
 			"operation": j.Type().Name,
@@ -89,7 +88,7 @@ func (j *taskExecutionTimeoutJob) Run(ctx context.Context) {
 			"host":      task.HostId,
 		}
 
-		if err := cleanUpTimedOutTask(task, displayTasksRestarted); err != nil {
+		if err := cleanUpTimedOutTask(task); err != nil {
 			grip.Warning(message.WrapError(err, msg))
 			j.AddError(err)
 			continue
@@ -106,7 +105,7 @@ func (j *taskExecutionTimeoutJob) Run(ctx context.Context) {
 }
 
 // function to clean up a single task
-func cleanUpTimedOutTask(t task.Task, displayTasksRestarted map[string]bool) error {
+func cleanUpTimedOutTask(t task.Task) error {
 	// get tlhe host for the task
 	host, err := host.FindOne(host.ById(t.HostId))
 	if err != nil {
@@ -141,17 +140,8 @@ func cleanUpTimedOutTask(t task.Task, displayTasksRestarted map[string]bool) err
 	}
 
 	// try to reset the task
-	taskID := t.Id
 	if t.IsPartOfDisplay() {
-		if displayTasksRestarted[taskID] {
-			return nil // already restarted this display task this run
-		}
-		taskID = t.DisplayTask.Id
-		displayTasksRestarted[taskID] = true
+		return t.DisplayTask.SetResetWhenFinished(detail)
 	}
-	if err := model.TryResetTask(taskID, "", "monitor", detail); err != nil {
-		return errors.Wrapf(err, "error trying to reset task %s", t.Id)
-	}
-
-	return nil
+	return errors.Wrapf(model.TryResetTask(t.Id, "", "monitor", detail), "error trying to reset task %s", t.Id)
 }
